+++ test/neighbor_test.rb
@@ -184,18 +184,6 @@ class NeighborTest < Minitest::Test
     load(file.path)
   end
 
-  def test_no_attribute
-    create_items(Item, :neighbor_vector)
-    result = Item.find(1).nearest_neighbors(distance: "euclidean").first(3)
-    assert_equal [3, 2], result.map(&:id)
-    assert_elements_in_delta [1, 1.7320507764816284], result.map(&:neighbor_distance)
-  end
-
-  def test_no_attribute_scope
-    create_items(Item, :neighbor_vector)
-    assert_equal [1], Item.nearest_neighbors([0, 0, 0], distance: "euclidean").limit(1).map(&:id)
-  end
-
   def test_invalid_attribute
     create_items(Item)
     error = assert_raises(ArgumentError) do
@@ -212,7 +200,7 @@ class NeighborTest < Minitest::Test
   end
 
   def test_neighbor_attributes
-    assert_equal Item.neighbor_attributes.keys, [:embedding, :neighbor_vector]
+    assert_equal Item.neighbor_attributes.keys, [:embedding]
   end
 
   def create_items(cls, attribute = :embedding)