+++ test/support/vector.rb
@@ -3,13 +3,11 @@ ActiveRecord::Schema.define do
 
   create_table :items, force: true do |t|
     t.vector :embedding, limit: 3
-    t.vector :neighbor_vector, limit: 3
   end
 end
 
 class Item < ActiveRecord::Base
   has_neighbors :embedding
-  has_neighbors
 end
 
 class CosineItem < ActiveRecord::Base