+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.3.0 (unreleased)
 
+- Removed default attribute name
 - Dropped support for Ruby < 3 and Active Record < 6.1
 
 ## 0.2.3 (2023-04-02)